Split "widgets" code out of ui::menu::global_state into a separate module
